Reverted r21374, and fixed the problem described therein by importing AITigerCompatibility.h, which defines NSInteger.

+++ Frameworks/AIUtilities Framework/Source/AIImageViewWithImagePicker.m
@@ -6,6 +6,8 @@
 //  Copyright (c) 2004-2005 The Adium Team. All rights reserved.
 //
 
+#import "AITigerCompatibility.h"
+
 #import "AIImageViewWithImagePicker.h"
 #import <Quartz/Quartz.h>
 
@@ -536,7 +538,7 @@
 	[self showPictureTaker];
 }
 
-- (void)pictureTakerDidEnd:(id)inPictureTaker returnCode:(int)returnCode contextInfo:(void *)contextInfo
+- (void)pictureTakerDidEnd:(id)inPictureTaker returnCode:(NSInteger)returnCode contextInfo:(void *)contextInfo
 {	
 	if (returnCode == NSOKButton) {
 		NSImage *image = [inPictureTaker outputImage];
